% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-lib.R
\name{addDuplicationsLines}
\alias{addDuplicationsLines}
\title{addDuplicationsLines}
\usage{
addDuplicationsLines(
  ggplot,
  x.min,
  x.max,
  linetype = "dotdash",
  line.color = "gray"
)
}
\description{
addDuplicationsLines
}
