% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{Countries}
\alias{Countries}
\title{New dataviz for reportGenerator by}
\description{
New dataviz for reportGenerator by

New dataviz for reportGenerator by
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Countries$new()}}
\item \href{#method-getContinentCountries}{\code{Countries$getContinentCountries()}}
\item \href{#method-clone}{\code{Countries$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Countries$new(countries)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getContinentCountries"></a>}}
\if{latex}{\out{\hypertarget{method-getContinentCountries}{}}}
\subsection{Method \code{getContinentCountries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Countries$getContinentCountries(cont)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Countries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
