% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataModel.R
\name{COVID19DataModel}
\alias{COVID19DataModel}
\title{COVID19DataModel}
\description{
COVID19DataModel

COVID19DataModel
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19DataModel$new()}}
\item \href{#method-addProblem}{\code{COVID19DataModel$addProblem()}}
\item \href{#method-loadData}{\code{COVID19DataModel$loadData()}}
\item \href{#method-checkField}{\code{COVID19DataModel$checkField()}}
\item \href{#method-getAggregatedData}{\code{COVID19DataModel$getAggregatedData()}}
\item \href{#method-clone}{\code{COVID19DataModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataModel$new(provider)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addProblem"></a>}}
\if{latex}{\out{\hypertarget{method-addProblem}{}}}
\subsection{Method \code{addProblem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataModel$addProblem(kind, key, value, obs = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataModel$loadData(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkField"></a>}}
\if{latex}{\out{\hypertarget{method-checkField}{}}}
\subsection{Method \code{checkField()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataModel$checkField(field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getAggregatedData"></a>}}
\if{latex}{\out{\hypertarget{method-getAggregatedData}{}}}
\subsection{Method \code{getAggregatedData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataModel$getAggregatedData(columns)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
