% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lib.R
\name{COVID19DataProcessor}
\alias{COVID19DataProcessor}
\title{COVID19DataProcessor}
\description{
COVID19DataProcessor

COVID19DataProcessor
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19DataProcessor$new()}}
\item \href{#method-curate}{\code{COVID19DataProcessor$curate()}}
\item \href{#method-downloadData}{\code{COVID19DataProcessor$downloadData()}}
\item \href{#method-loadData}{\code{COVID19DataProcessor$loadData()}}
\item \href{#method-cleanData}{\code{COVID19DataProcessor$cleanData()}}
\item \href{#method-consolidate}{\code{COVID19DataProcessor$consolidate()}}
\item \href{#method-makeDataComparation}{\code{COVID19DataProcessor$makeDataComparation()}}
\item \href{#method-makeImputationsOld}{\code{COVID19DataProcessor$makeImputationsOld()}}
\item \href{#method-makeImputations}{\code{COVID19DataProcessor$makeImputations()}}
\item \href{#method-makeImputationsNew2}{\code{COVID19DataProcessor$makeImputationsNew2()}}
\item \href{#method-calculateRates}{\code{COVID19DataProcessor$calculateRates()}}
\item \href{#method-calculateTopCountries}{\code{COVID19DataProcessor$calculateTopCountries()}}
\item \href{#method-clone}{\code{COVID19DataProcessor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$new(
  force.download = FALSE,
  imputation.method = ImputationMethodMean$new()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-curate"></a>}}
\subsection{Method \code{curate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$curate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadData"></a>}}
\subsection{Method \code{downloadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$downloadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanData"></a>}}
\subsection{Method \code{cleanData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$cleanData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidate"></a>}}
\subsection{Method \code{consolidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$consolidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeDataComparation"></a>}}
\subsection{Method \code{makeDataComparation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$makeDataComparation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeImputationsOld"></a>}}
\subsection{Method \code{makeImputationsOld()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$makeImputationsOld()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeImputations"></a>}}
\subsection{Method \code{makeImputations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$makeImputations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeImputationsNew2"></a>}}
\subsection{Method \code{makeImputationsNew2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$makeImputationsNew2()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateRates"></a>}}
\subsection{Method \code{calculateRates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$calculateRates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateTopCountries"></a>}}
\subsection{Method \code{calculateTopCountries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$calculateTopCountries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
