% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProcessor.R
\name{COVID19DataProviderConfirmedRecoveredDeaths}
\alias{COVID19DataProviderConfirmedRecoveredDeaths}
\title{COVID19DataProviderConfirmedRecoveredDeaths}
\description{
COVID19DataProviderConfirmedRecoveredDeaths

COVID19DataProviderConfirmedRecoveredDeaths
}
\author{
kenarab
}
\section{Super class}{
\code{\link[COVID19analytics:COVID19DataProvider]{COVID19analytics::COVID19DataProvider}} -> \code{COVID19DataProviderConfirmedRecoveredDeaths}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19DataProviderConfirmedRecoveredDeaths$new()}}
\item \href{#method-consolidate}{\code{COVID19DataProviderConfirmedRecoveredDeaths$consolidate()}}
\item \href{#method-downloadData}{\code{COVID19DataProviderConfirmedRecoveredDeaths$downloadData()}}
\item \href{#method-loadData}{\code{COVID19DataProviderConfirmedRecoveredDeaths$loadData()}}
\item \href{#method-cleanData}{\code{COVID19DataProviderConfirmedRecoveredDeaths$cleanData()}}
\item \href{#method-mergeData}{\code{COVID19DataProviderConfirmedRecoveredDeaths$mergeData()}}
\item \href{#method-clone}{\code{COVID19DataProviderConfirmedRecoveredDeaths$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="getDataModel">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-getDataModel}{\code{COVID19analytics::COVID19DataProvider$getDataModel()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="getDates">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-getDates}{\code{COVID19analytics::COVID19DataProvider$getDates()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="getID">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-getID}{\code{COVID19analytics::COVID19DataProvider$getID()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="setupData">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-setupData}{\code{COVID19analytics::COVID19DataProvider$setupData()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="standarize">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-standarize}{\code{COVID19analytics::COVID19DataProvider$standarize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderConfirmedRecoveredDeaths$new(force.download = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidate"></a>}}
\if{latex}{\out{\hypertarget{method-consolidate}{}}}
\subsection{Method \code{consolidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderConfirmedRecoveredDeaths$consolidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadData"></a>}}
\if{latex}{\out{\hypertarget{method-downloadData}{}}}
\subsection{Method \code{downloadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderConfirmedRecoveredDeaths$downloadData(
  download.freq = 60 * 60 * 18
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderConfirmedRecoveredDeaths$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanData"></a>}}
\if{latex}{\out{\hypertarget{method-cleanData}{}}}
\subsection{Method \code{cleanData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderConfirmedRecoveredDeaths$cleanData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mergeData"></a>}}
\if{latex}{\out{\hypertarget{method-mergeData}{}}}
\subsection{Method \code{mergeData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderConfirmedRecoveredDeaths$mergeData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderConfirmedRecoveredDeaths$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
