% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProcessor.R
\name{COVID19DataProviderCRD}
\alias{COVID19DataProviderCRD}
\title{COVID19DataProviderCRD
Data provider for Confirmed Recovered Deaths models separated in files}
\description{
COVID19DataProviderCRD
Data provider for Confirmed Recovered Deaths models separated in files

COVID19DataProviderCRD
Data provider for Confirmed Recovered Deaths models separated in files
}
\author{
kenarab
}
\section{Super class}{
\code{\link[COVID19analytics:COVID19DataProvider]{COVID19analytics::COVID19DataProvider}} -> \code{COVID19DataProviderCRD}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19DataProviderCRD$new()}}
\item \href{#method-consolidate}{\code{COVID19DataProviderCRD$consolidate()}}
\item \href{#method-downloadData}{\code{COVID19DataProviderCRD$downloadData()}}
\item \href{#method-loadData}{\code{COVID19DataProviderCRD$loadData()}}
\item \href{#method-cleanData}{\code{COVID19DataProviderCRD$cleanData()}}
\item \href{#method-mergeData}{\code{COVID19DataProviderCRD$mergeData()}}
\item \href{#method-clone}{\code{COVID19DataProviderCRD$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="getCitationInitials">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-getCitationInitials}{\code{COVID19analytics::COVID19DataProvider$getCitationInitials()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="getDataModel">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-getDataModel}{\code{COVID19analytics::COVID19DataProvider$getDataModel()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="getDates">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-getDates}{\code{COVID19analytics::COVID19DataProvider$getDates()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="getID">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-getID}{\code{COVID19analytics::COVID19DataProvider$getID()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="setupData">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-setupData}{\code{COVID19analytics::COVID19DataProvider$setupData()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="setupProcessor">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-setupProcessor}{\code{COVID19analytics::COVID19DataProvider$setupProcessor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="standarize">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-standarize}{\code{COVID19analytics::COVID19DataProvider$standarize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="transform">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-transform}{\code{COVID19analytics::COVID19DataProvider$transform()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderCRD$new(force.download = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidate"></a>}}
\if{latex}{\out{\hypertarget{method-consolidate}{}}}
\subsection{Method \code{consolidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderCRD$consolidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadData"></a>}}
\if{latex}{\out{\hypertarget{method-downloadData}{}}}
\subsection{Method \code{downloadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderCRD$downloadData(download.freq = 60 * 60 * 18)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderCRD$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanData"></a>}}
\if{latex}{\out{\hypertarget{method-cleanData}{}}}
\subsection{Method \code{cleanData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderCRD$cleanData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mergeData"></a>}}
\if{latex}{\out{\hypertarget{method-mergeData}{}}}
\subsection{Method \code{mergeData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderCRD$mergeData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderCRD$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
