% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProviderJHU.R
\name{COVID19DataProviderJHU}
\alias{COVID19DataProviderJHU}
\title{COVID19DataProviderJHU}
\description{
COVID19DataProviderJHU

COVID19DataProviderJHU
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19DataProviderJHU$new()}}
\item \href{#method-setupData}{\code{COVID19DataProviderJHU$setupData()}}
\item \href{#method-consolidate}{\code{COVID19DataProviderJHU$consolidate()}}
\item \href{#method-downloadData}{\code{COVID19DataProviderJHU$downloadData()}}
\item \href{#method-loadData}{\code{COVID19DataProviderJHU$loadData()}}
\item \href{#method-cleanData}{\code{COVID19DataProviderJHU$cleanData()}}
\item \href{#method-mergeData}{\code{COVID19DataProviderJHU$mergeData()}}
\item \href{#method-clone}{\code{COVID19DataProviderJHU$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderJHU$new(force.download = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupData"></a>}}
\if{latex}{\out{\hypertarget{method-setupData}{}}}
\subsection{Method \code{setupData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderJHU$setupData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidate"></a>}}
\if{latex}{\out{\hypertarget{method-consolidate}{}}}
\subsection{Method \code{consolidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderJHU$consolidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadData"></a>}}
\if{latex}{\out{\hypertarget{method-downloadData}{}}}
\subsection{Method \code{downloadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderJHU$downloadData(download.freq = 60 * 60 * 18)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderJHU$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanData"></a>}}
\if{latex}{\out{\hypertarget{method-cleanData}{}}}
\subsection{Method \code{cleanData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderJHU$cleanData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mergeData"></a>}}
\if{latex}{\out{\hypertarget{method-mergeData}{}}}
\subsection{Method \code{mergeData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderJHU$mergeData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderJHU$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
