% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProviderOWID.R
\name{COVID19DataProviderOWID}
\alias{COVID19DataProviderOWID}
\title{COVID19DataProviderOWID}
\description{
COVID19DataProviderOWID

COVID19DataProviderOWID
}
\author{
yanchang-zhao

kenarab
}
\section{Super class}{
\code{\link[COVID19analytics:COVID19DataProvider]{COVID19analytics::COVID19DataProvider}} -> \code{COVID19DataProviderOWID}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19DataProviderOWID$new()}}
\item \href{#method-getCitationInitials}{\code{COVID19DataProviderOWID$getCitationInitials()}}
\item \href{#method-getID}{\code{COVID19DataProviderOWID$getID()}}
\item \href{#method-downloadData}{\code{COVID19DataProviderOWID$downloadData()}}
\item \href{#method-getDates}{\code{COVID19DataProviderOWID$getDates()}}
\item \href{#method-loadData}{\code{COVID19DataProviderOWID$loadData()}}
\item \href{#method-consolidate}{\code{COVID19DataProviderOWID$consolidate()}}
\item \href{#method-cleanData}{\code{COVID19DataProviderOWID$cleanData()}}
\item \href{#method-mergeData}{\code{COVID19DataProviderOWID$mergeData()}}
\item \href{#method-standarize}{\code{COVID19DataProviderOWID$standarize()}}
\item \href{#method-clone}{\code{COVID19DataProviderOWID$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="getDataModel">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-getDataModel}{\code{COVID19analytics::COVID19DataProvider$getDataModel()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="setupData">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-setupData}{\code{COVID19analytics::COVID19DataProvider$setupData()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="setupProcessor">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-setupProcessor}{\code{COVID19analytics::COVID19DataProvider$setupProcessor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19DataProvider" data-id="transform">}\href{../../COVID19analytics/html/COVID19DataProvider.html#method-transform}{\code{COVID19analytics::COVID19DataProvider$transform()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$new(force.download = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCitationInitials"></a>}}
\if{latex}{\out{\hypertarget{method-getCitationInitials}{}}}
\subsection{Method \code{getCitationInitials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$getCitationInitials()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getID"></a>}}
\if{latex}{\out{\hypertarget{method-getID}{}}}
\subsection{Method \code{getID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$getID()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadData"></a>}}
\if{latex}{\out{\hypertarget{method-downloadData}{}}}
\subsection{Method \code{downloadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$downloadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDates"></a>}}
\if{latex}{\out{\hypertarget{method-getDates}{}}}
\subsection{Method \code{getDates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$getDates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidate"></a>}}
\if{latex}{\out{\hypertarget{method-consolidate}{}}}
\subsection{Method \code{consolidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$consolidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanData"></a>}}
\if{latex}{\out{\hypertarget{method-cleanData}{}}}
\subsection{Method \code{cleanData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$cleanData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mergeData"></a>}}
\if{latex}{\out{\hypertarget{method-mergeData}{}}}
\subsection{Method \code{mergeData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$mergeData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standarize"></a>}}
\if{latex}{\out{\hypertarget{method-standarize}{}}}
\subsection{Method \code{standarize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$standarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProviderOWID$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
