% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19MissingValuesImputation.R
\name{COVID19MissingValuesImputation}
\alias{COVID19MissingValuesImputation}
\title{COVID19MissingValuesImputation}
\description{
COVID19MissingValuesImputation

COVID19MissingValuesImputation
}
\author{
kenarab
}
\section{Super class}{
\code{\link[COVID19analytics:COVID19MissingValuesModel]{COVID19analytics::COVID19MissingValuesModel}} -> \code{COVID19MissingValuesImputation}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19MissingValuesImputation$new()}}
\item \href{#method-getID}{\code{COVID19MissingValuesImputation$getID()}}
\item \href{#method-apply}{\code{COVID19MissingValuesImputation$apply()}}
\item \href{#method-makeImputations}{\code{COVID19MissingValuesImputation$makeImputations()}}
\item \href{#method-clone}{\code{COVID19MissingValuesImputation$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="COVID19MissingValuesModel" data-id="setup">}\href{../../COVID19analytics/html/COVID19MissingValuesModel.html#method-setup}{\code{COVID19analytics::COVID19MissingValuesModel$setup()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19MissingValuesImputation$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getID"></a>}}
\if{latex}{\out{\hypertarget{method-getID}{}}}
\subsection{Method \code{getID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19MissingValuesImputation$getID()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply"></a>}}
\if{latex}{\out{\hypertarget{method-apply}{}}}
\subsection{Method \code{apply()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19MissingValuesImputation$apply()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeImputations"></a>}}
\if{latex}{\out{\hypertarget{method-makeImputations}{}}}
\subsection{Method \code{makeImputations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19MissingValuesImputation$makeImputations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19MissingValuesImputation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
