% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-lib.R
\name{COVID19TestCaseGenerator}
\alias{COVID19TestCaseGenerator}
\title{COVID19TestCaseGenerator}
\description{
COVID19TestCaseGenerator

COVID19TestCaseGenerator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19TestCaseGenerator$new()}}
\item \href{#method-generateTestCase}{\code{COVID19TestCaseGenerator$generateTestCase()}}
\item \href{#method-clone}{\code{COVID19TestCaseGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$new(
  data.processor,
  case.name,
  countries,
  test.case.folder = "inst/extdata/"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateTestCase"></a>}}
\if{latex}{\out{\hypertarget{method-generateTestCase}{}}}
\subsection{Method \code{generateTestCase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$generateTestCase(countries)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
