% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{downloadCOVID19}
\alias{downloadCOVID19}
\title{download COVID-19 data from Johns Hopkins University}
\usage{
downloadCOVID19(
  url.path,
  filename,
  force = FALSE,
  download.freq = 60 * 60 * 24,
  check.remote = FALSE,
  archive = TRUE
)
}
\description{
download COVID-19 data from Johns Hopkins University
}
\author{
kenarab
}
