% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProviderOWID.R
\name{downloadCOVID19OWID}
\alias{downloadCOVID19OWID}
\title{downloadCOVID19OWID download COVID-19 data from Johns Hopkins University}
\usage{
downloadCOVID19OWID(
  url.path,
  filename,
  force = FALSE,
  daily.update.time = "21:00:00",
  archive = TRUE
)
}
\description{
downloadCOVID19OWID download COVID-19 data from Johns Hopkins University
}
\author{
kenarab
}
