% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProviderJHU.R
\name{downloadJHUCOVID19}
\alias{downloadJHUCOVID19}
\title{download COVID-19 data from Johns Hopkins University}
\usage{
downloadJHUCOVID19(
  url.path,
  filename,
  force.download = FALSE,
  daily.update.time = "21:00:00",
  archive = TRUE
)
}
\description{
download COVID-19 data from Johns Hopkins University
}
\author{
kenarab
}
