% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{getEnv}
\alias{getEnv}
\title{getEnv}
\usage{
getEnv(
  variable.name,
  package.prefix = getPackagePrefix(),
  fail.on.empty = TRUE,
  env.file = "~/.Renviron",
  call.counter = 0
)
}
\description{
getEnv
}
\author{
kenarab
}
