% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19MissingValuesImputation.R
\name{ImputationMethod}
\alias{ImputationMethod}
\title{ImputationMethod}
\description{
ImputationMethod

ImputationMethod
}
\author{
ken4rab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ImputationMethod$new()}}
\item \href{#method-setup}{\code{ImputationMethod$setup()}}
\item \href{#method-getImputationValue}{\code{ImputationMethod$getImputationValue()}}
\item \href{#method-clone}{\code{ImputationMethod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethod$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethod$setup(data.processor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getImputationValue"></a>}}
\if{latex}{\out{\hypertarget{method-getImputationValue}{}}}
\subsection{Method \code{getImputationValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethod$getImputationValue(current.data, field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
