% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lib.R
\name{ImputationMethodMean}
\alias{ImputationMethodMean}
\title{ImputationMethodMean}
\description{
ImputationMethodMean

ImputationMethodMean
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ImputationMethodMean$new()}}
\item \href{#method-setup}{\code{ImputationMethodMean$setup()}}
\item \href{#method-getImputationValue}{\code{ImputationMethodMean$getImputationValue()}}
\item \href{#method-clone}{\code{ImputationMethodMean$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethodMean$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\subsection{Method \code{setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethodMean$setup(data.comparation)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getImputationValue"></a>}}
\subsection{Method \code{getImputationValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethodMean$getImputationValue(current.data, field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethodMean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
