% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{ReportGenerator}
\alias{ReportGenerator}
\title{ReportGenerator}
\description{
ReportGenerator

ReportGenerator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReportGenerator$new()}}
\item \href{#method-generateReport}{\code{ReportGenerator$generateReport()}}
\item \href{#method-downloadData}{\code{ReportGenerator$downloadData()}}
\item \href{#method-loadData}{\code{ReportGenerator$loadData()}}
\item \href{#method-cleanData}{\code{ReportGenerator$cleanData()}}
\item \href{#method-clone}{\code{ReportGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateReport"></a>}}
\subsection{Method \code{generateReport()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$generateReport(output.file, overwrite = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadData"></a>}}
\subsection{Method \code{downloadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$downloadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanData"></a>}}
\subsection{Method \code{cleanData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$cleanData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
