% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-lib.R
\name{ReportGeneratorDataComparison}
\alias{ReportGeneratorDataComparison}
\title{New dataviz for reportGenerator by}
\description{
New dataviz for reportGenerator by

New dataviz for reportGenerator by
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReportGeneratorDataComparison$new()}}
\item \href{#method-ggplotComparisonExponentialGrowth}{\code{ReportGeneratorDataComparison$ggplotComparisonExponentialGrowth()}}
\item \href{#method-getXLabelsTheme}{\code{ReportGeneratorDataComparison$getXLabelsTheme()}}
\item \href{#method-clone}{\code{ReportGeneratorDataComparison$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorDataComparison$new(data.processor, ma.n = 3)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotComparisonExponentialGrowth"></a>}}
\if{latex}{\out{\hypertarget{method-ggplotComparisonExponentialGrowth}{}}}
\subsection{Method \code{ggplotComparisonExponentialGrowth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorDataComparison$ggplotComparisonExponentialGrowth(
  included.countries,
  field = "confirmed",
  y.label = "Confirmed Cases",
  countries.text = "Top countries",
  min.cases = 100,
  show.legend = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getXLabelsTheme"></a>}}
\if{latex}{\out{\hypertarget{method-getXLabelsTheme}{}}}
\subsection{Method \code{getXLabelsTheme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorDataComparison$getXLabelsTheme(ggplot, x.values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorDataComparison$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
