% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-lib.R
\name{setupTheme}
\alias{setupTheme}
\title{setup Dataviz theme}
\usage{
setupTheme(
  ggplot,
  report.date,
  x.values,
  data.processor,
  total.colors,
  x.type = "dates",
  base.size = 6,
  log.scale.y = FALSE,
  log.scale.x = FALSE
)
}
\description{
setup Dataviz theme
}
