% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{smoothSerie2}
\alias{smoothSerie2}
\title{smooth a serie applying smooth package. Currently under construction}
\usage{
smoothSerie2(serie.name, serie, n)
}
\description{
smooth a serie applying smooth package. Currently under construction
}
\author{
kenarab
}
