% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-lib.R
\name{sourceRepoDiagnostic}
\alias{sourceRepoDiagnostic}
\title{Diagnostic update situation of source repository}
\usage{
sourceRepoDiagnostic(min.confirmed = 20)
}
\description{
Diagnostic update situation of source repository
}
\author{
kenarab
}
