% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{TexBuilder}
\alias{TexBuilder}
\title{TexBuilder}
\description{
TexBuilder

TexBuilder
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TexBuilder$new()}}
\item \href{#method-initTex}{\code{TexBuilder$initTex()}}
\item \href{#method-endTex}{\code{TexBuilder$endTex()}}
\item \href{#method-clone}{\code{TexBuilder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TexBuilder$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initTex"></a>}}
\subsection{Method \code{initTex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TexBuilder$initTex(output.file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-endTex"></a>}}
\subsection{Method \code{endTex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TexBuilder$endTex()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TexBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
