% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProviderJHU.R
\name{transformDataJHU}
\alias{transformDataJHU}
\title{Transform data from wide to long format and group by country}
\usage{
transformDataJHU(data)
}
\description{
Transform data from wide to long format and group by country
}
\author{
kenarab

yanchang-zhao
}
