% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProviderJHU.R
\name{transformDataJHUCountry}
\alias{transformDataJHUCountry}
\title{Transform data from wide to long format and group by country}
\usage{
transformDataJHUCountry(data)
}
\description{
Transform data from wide to long format and group by country
}
\author{
yanchang-zhao
}
