% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProviderOWID.R
\name{transformDataOWID}
\alias{transformDataOWID}
\title{Transform data from wide to long format and group by country}
\usage{
transformDataOWID(data)
}
\description{
Transform data from wide to long format and group by country
}
\author{
kenarab

yanchang-zhao
}
