% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProviderOWID.R
\name{transformDataOWIDCountry}
\alias{transformDataOWIDCountry}
\title{Transform data from wide to long format and group by country}
\usage{
transformDataOWIDCountry(data)
}
\description{
Transform data from wide to long format and group by country
}
\author{
yanchang-zhao
}
