% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lib.R
\name{COVID19DataProcessor}
\alias{COVID19DataProcessor}
\title{COVID19DataProcessor}
\description{
COVID19DataProcessor

COVID19DataProcessor
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19DataProcessor$new()}}
\item \href{#method-generateReport}{\code{COVID19DataProcessor$generateReport()}}
\item \href{#method-preprocess}{\code{COVID19DataProcessor$preprocess()}}
\item \href{#method-downloadData}{\code{COVID19DataProcessor$downloadData()}}
\item \href{#method-loadData}{\code{COVID19DataProcessor$loadData()}}
\item \href{#method-cleanData}{\code{COVID19DataProcessor$cleanData()}}
\item \href{#method-consolidate}{\code{COVID19DataProcessor$consolidate()}}
\item \href{#method-makeImputations}{\code{COVID19DataProcessor$makeImputations()}}
\item \href{#method-makeImputationsNew}{\code{COVID19DataProcessor$makeImputationsNew()}}
\item \href{#method-calculateRates}{\code{COVID19DataProcessor$calculateRates()}}
\item \href{#method-calculateTopCountries}{\code{COVID19DataProcessor$calculateTopCountries()}}
\item \href{#method-ggplotTopCountriesPie}{\code{COVID19DataProcessor$ggplotTopCountriesPie()}}
\item \href{#method-ggplotTopCountriesBarPlots}{\code{COVID19DataProcessor$ggplotTopCountriesBarPlots()}}
\item \href{#method-ggplotCountriesBarGraphs}{\code{COVID19DataProcessor$ggplotCountriesBarGraphs()}}
\item \href{#method-ggplotConfirmedCases}{\code{COVID19DataProcessor$ggplotConfirmedCases()}}
\item \href{#method-generateTex}{\code{COVID19DataProcessor$generateTex()}}
\item \href{#method-getXLabelsTheme}{\code{COVID19DataProcessor$getXLabelsTheme()}}
\item \href{#method-clone}{\code{COVID19DataProcessor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$new(force.download = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateReport"></a>}}
\subsection{Method \code{generateReport()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$generateReport(output.file, overwrite = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocess"></a>}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$preprocess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadData"></a>}}
\subsection{Method \code{downloadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$downloadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanData"></a>}}
\subsection{Method \code{cleanData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$cleanData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidate"></a>}}
\subsection{Method \code{consolidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$consolidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeImputations"></a>}}
\subsection{Method \code{makeImputations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$makeImputations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeImputationsNew"></a>}}
\subsection{Method \code{makeImputationsNew()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$makeImputationsNew()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateRates"></a>}}
\subsection{Method \code{calculateRates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$calculateRates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateTopCountries"></a>}}
\subsection{Method \code{calculateTopCountries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$calculateTopCountries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotTopCountriesPie"></a>}}
\subsection{Method \code{ggplotTopCountriesPie()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$ggplotTopCountriesPie(excluded.countries = "World")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotTopCountriesBarPlots"></a>}}
\subsection{Method \code{ggplotTopCountriesBarPlots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$ggplotTopCountriesBarPlots(excluded.countries = "World")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotCountriesBarGraphs"></a>}}
\subsection{Method \code{ggplotCountriesBarGraphs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$ggplotCountriesBarGraphs(selected.country = "Australia")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotConfirmedCases"></a>}}
\subsection{Method \code{ggplotConfirmedCases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$ggplotConfirmedCases()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateTex"></a>}}
\subsection{Method \code{generateTex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$generateTex(output.file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getXLabelsTheme"></a>}}
\subsection{Method \code{getXLabelsTheme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$getXLabelsTheme(ggplot, x.values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
