% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProcessor.R
\name{COVID19DataProcessor}
\alias{COVID19DataProcessor}
\title{COVID19DataProcessor}
\description{
COVID19DataProcessor

COVID19DataProcessor
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19DataProcessor$new()}}
\item \href{#method-initProviders}{\code{COVID19DataProcessor$initProviders()}}
\item \href{#method-getDataProvider}{\code{COVID19DataProcessor$getDataProvider()}}
\item \href{#method-initMissingValuesModels}{\code{COVID19DataProcessor$initMissingValuesModels()}}
\item \href{#method-getData}{\code{COVID19DataProcessor$getData()}}
\item \href{#method-setupProvider}{\code{COVID19DataProcessor$setupProvider()}}
\item \href{#method-setupMissingValuesModel}{\code{COVID19DataProcessor$setupMissingValuesModel()}}
\item \href{#method-setupStrategies}{\code{COVID19DataProcessor$setupStrategies()}}
\item \href{#method-setupProcessor}{\code{COVID19DataProcessor$setupProcessor()}}
\item \href{#method-setupData}{\code{COVID19DataProcessor$setupData()}}
\item \href{#method-process}{\code{COVID19DataProcessor$process()}}
\item \href{#method-transform}{\code{COVID19DataProcessor$transform()}}
\item \href{#method-checkValidTransition}{\code{COVID19DataProcessor$checkValidTransition()}}
\item \href{#method-changeState}{\code{COVID19DataProcessor$changeState()}}
\item \href{#method-curate}{\code{COVID19DataProcessor$curate()}}
\item \href{#method-makeDataComparison}{\code{COVID19DataProcessor$makeDataComparison()}}
\item \href{#method-getCountries}{\code{COVID19DataProcessor$getCountries()}}
\item \href{#method-smoothSeries}{\code{COVID19DataProcessor$smoothSeries()}}
\item \href{#method-calculateRates}{\code{COVID19DataProcessor$calculateRates()}}
\item \href{#method-calculateTopCountries}{\code{COVID19DataProcessor$calculateTopCountries()}}
\item \href{#method-clone}{\code{COVID19DataProcessor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$new(
  provider.id,
  missing.values.model.id,
  force.download = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initProviders"></a>}}
\if{latex}{\out{\hypertarget{method-initProviders}{}}}
\subsection{Method \code{initProviders()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$initProviders()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDataProvider"></a>}}
\if{latex}{\out{\hypertarget{method-getDataProvider}{}}}
\subsection{Method \code{getDataProvider()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$getDataProvider()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initMissingValuesModels"></a>}}
\if{latex}{\out{\hypertarget{method-initMissingValuesModels}{}}}
\subsection{Method \code{initMissingValuesModels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$initMissingValuesModels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getData"></a>}}
\if{latex}{\out{\hypertarget{method-getData}{}}}
\subsection{Method \code{getData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$getData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupProvider"></a>}}
\if{latex}{\out{\hypertarget{method-setupProvider}{}}}
\subsection{Method \code{setupProvider()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$setupProvider(provider.id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupMissingValuesModel"></a>}}
\if{latex}{\out{\hypertarget{method-setupMissingValuesModel}{}}}
\subsection{Method \code{setupMissingValuesModel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$setupMissingValuesModel(missing.values.model.id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupStrategies"></a>}}
\if{latex}{\out{\hypertarget{method-setupStrategies}{}}}
\subsection{Method \code{setupStrategies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$setupStrategies()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupProcessor"></a>}}
\if{latex}{\out{\hypertarget{method-setupProcessor}{}}}
\subsection{Method \code{setupProcessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$setupProcessor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupData"></a>}}
\if{latex}{\out{\hypertarget{method-setupData}{}}}
\subsection{Method \code{setupData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$setupData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-process"></a>}}
\if{latex}{\out{\hypertarget{method-process}{}}}
\subsection{Method \code{process()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$process()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$transform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-checkValidTransition"></a>}}
\if{latex}{\out{\hypertarget{method-checkValidTransition}{}}}
\subsection{Method \code{checkValidTransition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$checkValidTransition(
  state.expected = "datamodel-setup",
  fail.on.error = TRUE,
  only.check = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-changeState"></a>}}
\if{latex}{\out{\hypertarget{method-changeState}{}}}
\subsection{Method \code{changeState()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$changeState(new.state)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-curate"></a>}}
\if{latex}{\out{\hypertarget{method-curate}{}}}
\subsection{Method \code{curate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$curate(countries = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeDataComparison"></a>}}
\if{latex}{\out{\hypertarget{method-makeDataComparison}{}}}
\subsection{Method \code{makeDataComparison()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$makeDataComparison()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCountries"></a>}}
\if{latex}{\out{\hypertarget{method-getCountries}{}}}
\subsection{Method \code{getCountries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$getCountries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-smoothSeries"></a>}}
\if{latex}{\out{\hypertarget{method-smoothSeries}{}}}
\subsection{Method \code{smoothSeries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$smoothSeries(old.serie.sufix = "original")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateRates"></a>}}
\if{latex}{\out{\hypertarget{method-calculateRates}{}}}
\subsection{Method \code{calculateRates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$calculateRates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculateTopCountries"></a>}}
\if{latex}{\out{\hypertarget{method-calculateTopCountries}{}}}
\subsection{Method \code{calculateTopCountries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$calculateTopCountries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
