% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProcessor.R
\name{COVID19DataProvider}
\alias{COVID19DataProvider}
\title{COVID19DataProvider}
\description{
COVID19DataProvider

COVID19DataProvider
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19DataProvider$new()}}
\item \href{#method-getDataModel}{\code{COVID19DataProvider$getDataModel()}}
\item \href{#method-getID}{\code{COVID19DataProvider$getID()}}
\item \href{#method-getDates}{\code{COVID19DataProvider$getDates()}}
\item \href{#method-transform}{\code{COVID19DataProvider$transform()}}
\item \href{#method-setupProcessor}{\code{COVID19DataProvider$setupProcessor()}}
\item \href{#method-setupData}{\code{COVID19DataProvider$setupData()}}
\item \href{#method-downloadData}{\code{COVID19DataProvider$downloadData()}}
\item \href{#method-getCitationInitials}{\code{COVID19DataProvider$getCitationInitials()}}
\item \href{#method-loadData}{\code{COVID19DataProvider$loadData()}}
\item \href{#method-cleanData}{\code{COVID19DataProvider$cleanData()}}
\item \href{#method-consolidate}{\code{COVID19DataProvider$consolidate()}}
\item \href{#method-standarize}{\code{COVID19DataProvider$standarize()}}
\item \href{#method-clone}{\code{COVID19DataProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$new(force.download = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDataModel"></a>}}
\if{latex}{\out{\hypertarget{method-getDataModel}{}}}
\subsection{Method \code{getDataModel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$getDataModel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getID"></a>}}
\if{latex}{\out{\hypertarget{method-getID}{}}}
\subsection{Method \code{getID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$getID()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDates"></a>}}
\if{latex}{\out{\hypertarget{method-getDates}{}}}
\subsection{Method \code{getDates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$getDates()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$transform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupProcessor"></a>}}
\if{latex}{\out{\hypertarget{method-setupProcessor}{}}}
\subsection{Method \code{setupProcessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$setupProcessor(data.processor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupData"></a>}}
\if{latex}{\out{\hypertarget{method-setupData}{}}}
\subsection{Method \code{setupData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$setupData(data.processor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-downloadData"></a>}}
\if{latex}{\out{\hypertarget{method-downloadData}{}}}
\subsection{Method \code{downloadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$downloadData(download.freq = 60 * 60 * 18)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCitationInitials"></a>}}
\if{latex}{\out{\hypertarget{method-getCitationInitials}{}}}
\subsection{Method \code{getCitationInitials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$getCitationInitials()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadData"></a>}}
\if{latex}{\out{\hypertarget{method-loadData}{}}}
\subsection{Method \code{loadData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$loadData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cleanData"></a>}}
\if{latex}{\out{\hypertarget{method-cleanData}{}}}
\subsection{Method \code{cleanData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$cleanData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-consolidate"></a>}}
\if{latex}{\out{\hypertarget{method-consolidate}{}}}
\subsection{Method \code{consolidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$consolidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-standarize"></a>}}
\if{latex}{\out{\hypertarget{method-standarize}{}}}
\subsection{Method \code{standarize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$standarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19DataProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
