% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19DataProcessor.R
\name{COVID19MissingValuesModel}
\alias{COVID19MissingValuesModel}
\title{COVID19MissingValuesModel}
\description{
COVID19MissingValuesModel

COVID19MissingValuesModel
}
\author{
kenarab
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19MissingValuesModel$new()}}
\item \href{#method-getID}{\code{COVID19MissingValuesModel$getID()}}
\item \href{#method-setup}{\code{COVID19MissingValuesModel$setup()}}
\item \href{#method-apply}{\code{COVID19MissingValuesModel$apply()}}
\item \href{#method-clone}{\code{COVID19MissingValuesModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19MissingValuesModel$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getID"></a>}}
\if{latex}{\out{\hypertarget{method-getID}{}}}
\subsection{Method \code{getID()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19MissingValuesModel$getID()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19MissingValuesModel$setup(data.processor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-apply"></a>}}
\if{latex}{\out{\hypertarget{method-apply}{}}}
\subsection{Method \code{apply()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19MissingValuesModel$apply()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19MissingValuesModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
