% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-lib.R
\name{COVID19TestCaseGenerator}
\alias{COVID19TestCaseGenerator}
\title{COVID19TestCaseGenerator}
\description{
COVID19TestCaseGenerator

COVID19TestCaseGenerator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{COVID19TestCaseGenerator$new()}}
\item \href{#method-generateTestCase}{\code{COVID19TestCaseGenerator$generateTestCase()}}
\item \href{#method-getSourceIndicatorPath}{\code{COVID19TestCaseGenerator$getSourceIndicatorPath()}}
\item \href{#method-getExpectedFile}{\code{COVID19TestCaseGenerator$getExpectedFile()}}
\item \href{#method-readExpectedFile}{\code{COVID19TestCaseGenerator$readExpectedFile()}}
\item \href{#method-getProcessor}{\code{COVID19TestCaseGenerator$getProcessor()}}
\item \href{#method-getProcessorPreloaded}{\code{COVID19TestCaseGenerator$getProcessorPreloaded()}}
\item \href{#method-doRegressionTest}{\code{COVID19TestCaseGenerator$doRegressionTest()}}
\item \href{#method-clone}{\code{COVID19TestCaseGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$new(
  case.name,
  countries,
  provider.id = "JohnsHopkingsUniversity",
  missing.values.model.id = "imputation",
  running.test = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateTestCase"></a>}}
\if{latex}{\out{\hypertarget{method-generateTestCase}{}}}
\subsection{Method \code{generateTestCase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$generateTestCase()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSourceIndicatorPath"></a>}}
\if{latex}{\out{\hypertarget{method-getSourceIndicatorPath}{}}}
\subsection{Method \code{getSourceIndicatorPath()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$getSourceIndicatorPath(indicator)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExpectedFile"></a>}}
\if{latex}{\out{\hypertarget{method-getExpectedFile}{}}}
\subsection{Method \code{getExpectedFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$getExpectedFile()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-readExpectedFile"></a>}}
\if{latex}{\out{\hypertarget{method-readExpectedFile}{}}}
\subsection{Method \code{readExpectedFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$readExpectedFile(expected.file.path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProcessor"></a>}}
\if{latex}{\out{\hypertarget{method-getProcessor}{}}}
\subsection{Method \code{getProcessor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$getProcessor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getProcessorPreloaded"></a>}}
\if{latex}{\out{\hypertarget{method-getProcessorPreloaded}{}}}
\subsection{Method \code{getProcessorPreloaded()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$getProcessorPreloaded()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doRegressionTest"></a>}}
\if{latex}{\out{\hypertarget{method-doRegressionTest}{}}}
\subsection{Method \code{doRegressionTest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$doRegressionTest(rownum2test, seed = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{COVID19TestCaseGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
