% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lib.R
\name{ImputationMethodMean}
\alias{ImputationMethodMean}
\title{ImputationMethodMean}
\description{
ImputationMethodMean

ImputationMethodMean
}
\author{
ken4rab
}
\section{Super class}{
\code{\link[COVID19:ImputationMethod]{COVID19::ImputationMethod}} -> \code{ImputationMethodMean}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ImputationMethodMean$new()}}
\item \href{#method-getImputationValue}{\code{ImputationMethodMean$getImputationValue()}}
\item \href{#method-clone}{\code{ImputationMethodMean$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="COVID19" data-topic="ImputationMethod" data-id="setup">}\href{../../COVID19/html/ImputationMethod.html#method-setup}{\code{COVID19::ImputationMethod$setup()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethodMean$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getImputationValue"></a>}}
\if{latex}{\out{\hypertarget{method-getImputationValue}{}}}
\subsection{Method \code{getImputationValue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethodMean$getImputationValue(current.data, field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImputationMethodMean$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
