% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-lib.R
\name{ReportGenerator}
\alias{ReportGenerator}
\title{ReportGenerator}
\description{
ReportGenerator

ReportGenerator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReportGenerator$new()}}
\item \href{#method-preprocess}{\code{ReportGenerator$preprocess()}}
\item \href{#method-generateReport}{\code{ReportGenerator$generateReport()}}
\item \href{#method-ggplotTopCountriesPie}{\code{ReportGenerator$ggplotTopCountriesPie()}}
\item \href{#method-ggplotTopCountriesBarPlots}{\code{ReportGenerator$ggplotTopCountriesBarPlots()}}
\item \href{#method-ggplotCountriesBarGraphs}{\code{ReportGenerator$ggplotCountriesBarGraphs()}}
\item \href{#method-ggplotConfirmedCases}{\code{ReportGenerator$ggplotConfirmedCases()}}
\item \href{#method-generateTex}{\code{ReportGenerator$generateTex()}}
\item \href{#method-getXLabelsTheme}{\code{ReportGenerator$getXLabelsTheme()}}
\item \href{#method-clone}{\code{ReportGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$new(data.processor)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preprocess"></a>}}
\if{latex}{\out{\hypertarget{method-preprocess}{}}}
\subsection{Method \code{preprocess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$preprocess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateReport"></a>}}
\if{latex}{\out{\hypertarget{method-generateReport}{}}}
\subsection{Method \code{generateReport()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$generateReport(output.file, overwrite = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotTopCountriesPie"></a>}}
\if{latex}{\out{\hypertarget{method-ggplotTopCountriesPie}{}}}
\subsection{Method \code{ggplotTopCountriesPie()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$ggplotTopCountriesPie(excluded.countries = "World")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotTopCountriesBarPlots"></a>}}
\if{latex}{\out{\hypertarget{method-ggplotTopCountriesBarPlots}{}}}
\subsection{Method \code{ggplotTopCountriesBarPlots()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$ggplotTopCountriesBarPlots(excluded.countries = "World")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotCountriesBarGraphs"></a>}}
\if{latex}{\out{\hypertarget{method-ggplotCountriesBarGraphs}{}}}
\subsection{Method \code{ggplotCountriesBarGraphs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$ggplotCountriesBarGraphs(selected.country = "Australia")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotConfirmedCases"></a>}}
\if{latex}{\out{\hypertarget{method-ggplotConfirmedCases}{}}}
\subsection{Method \code{ggplotConfirmedCases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$ggplotConfirmedCases()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generateTex"></a>}}
\if{latex}{\out{\hypertarget{method-generateTex}{}}}
\subsection{Method \code{generateTex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$generateTex(output.file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getXLabelsTheme"></a>}}
\if{latex}{\out{\hypertarget{method-getXLabelsTheme}{}}}
\subsection{Method \code{getXLabelsTheme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$getXLabelsTheme(ggplot, x.values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
