% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-lib.R
\name{ReportGeneratorEnhanced}
\alias{ReportGeneratorEnhanced}
\title{ReportGeneratorEnhanced
New dataviz for reportGenerator by}
\description{
ReportGeneratorEnhanced
New dataviz for reportGenerator by

ReportGeneratorEnhanced
New dataviz for reportGenerator by
}
\author{
kenarab
}
\section{Super class}{
\code{\link[COVID19analytics:ReportGenerator]{COVID19analytics::ReportGenerator}} -> \code{ReportGeneratorEnhanced}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReportGeneratorEnhanced$new()}}
\item \href{#method-ggplotTopCountriesStackedBarDailyInc}{\code{ReportGeneratorEnhanced$ggplotTopCountriesStackedBarDailyInc()}}
\item \href{#method-ggplotCountriesLines}{\code{ReportGeneratorEnhanced$ggplotCountriesLines()}}
\item \href{#method-ggplotCrossSection}{\code{ReportGeneratorEnhanced$ggplotCrossSection()}}
\item \href{#method-clone}{\code{ReportGeneratorEnhanced$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="ReportGenerator" data-id="generateReport">}\href{../../COVID19analytics/html/ReportGenerator.html#method-generateReport}{\code{COVID19analytics::ReportGenerator$generateReport()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="ReportGenerator" data-id="generateTex">}\href{../../COVID19analytics/html/ReportGenerator.html#method-generateTex}{\code{COVID19analytics::ReportGenerator$generateTex()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="ReportGenerator" data-id="getXLabelsTheme">}\href{../../COVID19analytics/html/ReportGenerator.html#method-getXLabelsTheme}{\code{COVID19analytics::ReportGenerator$getXLabelsTheme()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="ReportGenerator" data-id="ggplotConfirmedCases">}\href{../../COVID19analytics/html/ReportGenerator.html#method-ggplotConfirmedCases}{\code{COVID19analytics::ReportGenerator$ggplotConfirmedCases()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="ReportGenerator" data-id="ggplotCountriesBarGraphs">}\href{../../COVID19analytics/html/ReportGenerator.html#method-ggplotCountriesBarGraphs}{\code{COVID19analytics::ReportGenerator$ggplotCountriesBarGraphs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="ReportGenerator" data-id="ggplotTopCountriesBarPlots">}\href{../../COVID19analytics/html/ReportGenerator.html#method-ggplotTopCountriesBarPlots}{\code{COVID19analytics::ReportGenerator$ggplotTopCountriesBarPlots()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="ReportGenerator" data-id="ggplotTopCountriesPie">}\href{../../COVID19analytics/html/ReportGenerator.html#method-ggplotTopCountriesPie}{\code{COVID19analytics::ReportGenerator$ggplotTopCountriesPie()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="COVID19analytics" data-topic="ReportGenerator" data-id="preprocess">}\href{../../COVID19analytics/html/ReportGenerator.html#method-preprocess}{\code{COVID19analytics::ReportGenerator$preprocess()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorEnhanced$new(data.processor, ma.n = 7)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotTopCountriesStackedBarDailyInc"></a>}}
\if{latex}{\out{\hypertarget{method-ggplotTopCountriesStackedBarDailyInc}{}}}
\subsection{Method \code{ggplotTopCountriesStackedBarDailyInc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorEnhanced$ggplotTopCountriesStackedBarDailyInc(
  included.countries,
  excluded.countries = "World",
  countries.text = "Top countries"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotCountriesLines"></a>}}
\if{latex}{\out{\hypertarget{method-ggplotCountriesLines}{}}}
\subsection{Method \code{ggplotCountriesLines()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorEnhanced$ggplotCountriesLines(
  included.countries = self$data.processor$top.countries,
  countries.text = "top countries",
  excluded.countries = "World",
  field = "confirmed.inc",
  field.description = "Daily new Confirmed Cases",
  log.scale = FALSE,
  show.legend = FALSE,
  min.confirmed = 100
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ggplotCrossSection"></a>}}
\if{latex}{\out{\hypertarget{method-ggplotCrossSection}{}}}
\subsection{Method \code{ggplotCrossSection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorEnhanced$ggplotCrossSection(
  included.countries = self$data.processor$top.countries,
  countries.text = "top countries",
  excluded.countries = "World",
  field.x = "confirmed",
  label.x = field.x,
  field.y = "death.rate.min",
  label.y = field.y,
  plot.description = "Cross section Confirmed vs  Death rate min",
  log.scale.x = TRUE,
  log.scale.y = FALSE,
  min.confirmed = 100,
  show.legend = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReportGeneratorEnhanced$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
